requirements.txt
llama-index
langchain
langchain-community
sentence-transformers
python-dotenv
llama-index-llms-fireworks 
ipykernel  
huggingface-hub  
streamlit
llama-index-embeddings-langchain
llama-index-llms-langchain
PyPdf2